requirements.txt
rasa==3.6.13
rasa-sdk==3.6.2
torch>=1.13.0
transformers==4.37.2
numpy
tqdm
qdrant-client
google-generativeai
python-dotenv